+++ e3nn_jax/__init__.py
@@ -38,7 +38,7 @@ from e3nn_jax._src.irreps import Irrep, MulIrrep, Irreps
 from e3nn_jax._src.irreps_array import IrrepsArray, concatenate, stack, mean, norm, normal
 from e3nn_jax._src.irreps_array import sum_ as sum
 from e3nn_jax._src.spherical_harmonics import spherical_harmonics, sh, legendre
-from e3nn_jax._src.radial import sus, soft_one_hot_linspace, bessel, poly_envelope
+from e3nn_jax._src.radial import sus, soft_one_hot_linspace, bessel, poly_envelope, soft_envelope
 from e3nn_jax._src.linear import FunctionalLinear, Linear
 from e3nn_jax._src.core_tensor_product import FunctionalTensorProduct
 from e3nn_jax._src.tensor_products import (
@@ -132,6 +132,7 @@ __all__ = [
     "index_add",
     "radius_graph",
     "poly_envelope",
+    "soft_envelope",
     "reduced_tensor_product_basis",
     "reduced_symmetric_tensor_product_basis",
     "SymmetricTensorProduct",