+++ ChangeLog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `e3nn.SymmetricTensorProduct` operation: a parameterized version of `x + x^2 + x^3 + ...`.
+- `e3nn.soft_envelope` a smooth `C^inf` envelope radial function.
 
 ## [0.10.0] - 2022-10-05
 ### Added